requirements.txt
pylint==2.17.7
python-dotenv==0.21.1
yapf==0.40.2
pandas==1.3.5